% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{box_muller}
\alias{box_muller}
\title{Box-muller}
\usage{
box_muller(mu, q)
}
\arguments{
\item{mu}{is mean of the distribution, one-dimensional.}

\item{q}{is precision of the distribution, one-dimensional.}
}
\description{
Drawing univariate Gaussian random variables with mean mu in R and precision q > 0.
The function transforms a pair of independent uniform random variables into a pair
of Gaussian random variables by exploiting the radial symmetry of the
two-dimensional normal distribution.
}
\examples{
box_muller(0,1)
}
