% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSAMS}
\alias{GSAMS}
\title{MCMC sampler based on approximate matrix splitting}
\usage{
GSAMS(mu, Q, total_itr, theta0, omega, method)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is arbitrary precision matrix.}

\item{total_itr}{is the total itration of Gibbs sampler}

\item{theta0}{is initialization.}

\item{method}{is matrix splitting approach to choose within "Clone" and "Hogwild".}
}
\description{
MCMC samplers whose invariant distributions are approximations
}
\examples{
mu <- 1:10
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GSAMS(mu,Q,500,rep(0,10),1, "Hogwild")
}
