% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSband}
\alias{GSband}
\title{Multivariate Gaussian sampling with sparse or band matrix Q}
\usage{
GSband(mu, Q, b)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is band precision matrix.}

\item{b}{is bandwidth}
}
\description{
To generate a d-dimensional Gaussian vector theta with mean mu and band precision matrix Q
}
\examples{
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GSband(rep(1,10),Q,3)
}
