% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GScheby}
\alias{GScheby}
\title{Chebyshev Polynomials Sampler}
\usage{
GScheby(mu, Q, K)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is arbitrary precision matrix.}

\item{K}{is the order of the Chebyshev series.}
}
\description{
Approx. square root sampler using Chebyshev polynomials
}
\examples{
mu <- 1:10
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GScheby(mu, Q, 10)
}
