% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSchol}
\alias{GSchol}
\title{Cholesky Factorization Sampler}
\usage{
GSchol(mu, Q)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is arbitrary precision matrix.}
}
\description{
Sampling based on the Cholesky decomposition on Q.
}
\examples{
mu <- 1:10
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GSchol(mu, Q)
}
