% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GScirculant}
\alias{GScirculant}
\title{Multivariate Gaussian sampling with block circulant (Toeplitz) matrix Q}
\usage{
GScirculant(mu, a, N, M)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{a}{is vector built by stacking the first columns associated to the Q.}

\item{N}{is dimension of each block.}

\item{M}{is number of different blocks.}
}
\description{
Q is a block circulant matrix, then the sampling can be performed in the Fourier domain.
}
\examples{
mu=rep(0,2); a=matrix(c(1,0),2,1); M=1; N=2;
GScirculant(mu, a, N, M)
}
