% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSdiag}
\alias{GSdiag}
\title{Multivariate Gaussian sampling with diagonal precision matrix}
\usage{
GSdiag(mu, Q)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is diagonal precision matrix.}
}
\description{
To generate a d-dimensional Gaussian vector theta with mean mu and diagonal precision matrix Q
}
\examples{
GSdiag((rep(1,5),diag(2,5,5)))
}
