% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSgibbs}
\alias{GSgibbs}
\title{Component-wise Gibbs sampler}
\usage{
GSgibbs(mu, Q, total_itr, theta0)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is arbitrary precision matrix.}

\item{total_itr}{is the total itration of Gibbs sampler}

\item{theta0}{is initialization.}
}
\description{
To sequentially draw one component of  given the others.
}
\examples{
mu <- 1:10
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GSgibbs(mu,Q,1000,rep(0,10))
}
