% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSlanczos}
\alias{GSlanczos}
\title{Lanczos Decomposition Sampler}
\usage{
GSlanczos(mu, Q, K)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is arbitrary precision matrix.}

\item{K}{is the dimension of Krylov subspace}
}
\description{
Approx. square root sampler using Lanczos decomposition
}
\examples{
mu <- 1:10
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GSlanczos(mu, Q, 10)
}
