% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSMS}
\alias{GSMS}
\title{MCMC sampler based on exact matrix splitting}
\usage{
GSMS(mu, Q, total_itr, theta0, method)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is arbitrary precision matrix.}

\item{total_itr}{is the total itration of Gibbs sampler}

\item{theta0}{is initialization.}

\item{method}{is matrix splitting approach to choose within "Gauss-Seidel", "Richardson", "Jacobi", "SOR".}
}
\description{
A perturbed instance of Matrix Splitting schemes which are a class of linear iterative solvers based on the splitting of Q into Q = M - N.
}
\examples{
mu <- 1:10
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GSMS(mu,Q,100,rep(0,10), "SOR")
}
