% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSMS_ssor}
\alias{GSMS_ssor}
\title{Chebyshev accelerated SSOR sampler}
\usage{
GSMS_ssor(mu, Q, total_itr, theta0)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is arbitrary precision matrix.}

\item{total_itr}{is the total itration of Gibbs sampler}

\item{theta0}{is initialization.}
}
\description{
an accelerated version of exact matrix splitting sampler based on the SSOR splitting
}
\examples{
mu <- 1:10
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GSMS_ssor(mu,Q,100,rep(0,10))
}
