% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GSsqrt}
\alias{GSsqrt}
\title{Square Root Factorization Sampler}
\usage{
GSsqrt(mu, Q)
}
\arguments{
\item{mu}{is mean, d-dimensional.}

\item{Q}{is arbitrary precision matrix.}
}
\description{
Useless
}
\examples{
mu <- 1:10
Q <- diag(2,10,10); Q[4,1]<-Q[1,4]<-1;
GSsqrt(mu, Q)
}
